M2-4827: Post password reset redirection (#373)
### 📝 Description

This PR redirects the user to the login page after successful password redirection and displays a success banner on that page. I've also removed the `backToLogin` button, and the form success message.

I'm using `useRef` as an escape hatch in the `LoginPage` component to prevent the banner from showing twice in strict mode. This is arguably a hack and should be generalised, but I didn't think it should be part of this ticket