M2-5533: Mobile deep linking files (#406)
This PR adds two JSON files to the respondent web app which is part of the requirement for supporting universal links on iOS, and app links on Android. Basically this means that if a user clicks on a link for the respondent web app and they have the app installed, it will open the app instead.

This initial support will only work for the password recovery page (see https://github.com/ChildMindInstitute/mindlogger-app-refactor/pull/665). Later on support will be added for more pages.