M2-5923: Create Update Jira Tickets GitHub Actions workflow (#422)
This PR introduces a GitHub Actions workflow that submits a list of ticket numbers to a Jira webhook based on a number of factors.

The workflow is triggered by a GitHub release, and it collates the list of Jira ticket numbers that were worked on since the last GitHub release that had a successful Jenkins build. It then waits for the outcome of the current build (trigged by the release) and submits that list of Jira tickets if the build is successful.

My original approach to this ticket was to modify the Jenkins scripts, but I changed that because it's much simpler to have this logic in one workflow (instead of being split across multiple CodeCommit repos), and we eventually intent to move CI/CD to GitHub Actions anyway so this is a step in that direction

(cherry picked from commit 1c04959442be5906880201a82724e0f255bd1165)