M2-5923: Fix release event type (#425)
This PR follows up https://github.com/ChildMindInstitute/mindlogger-web-refactor/pull/422 by changing the release event type to `published` instead of `created`. The `created` event type doesn't capture releases that were previously saved as draft, so some releases would be missed.

This change also brings the workflow file to the dev branch. This is important because that's where the tagging takes place and the workflow file won't be triggered if it doesn't exist on that branch.