M2-6212: Allow direct linking to activity (#431)
This change allows directly linking to an activity by wrapping the `< AppletDetailsPage />` component in an `<AuthorizationGuard />`. I also created a new reusable component called `<LoginWithRedirect />` that, in combination with the `<AuthorizationGuard />`, automatically takes the user to the login page if they try to access a page while unauthenticated, and then redirects them to wherever they originally tried to access.

I also removed the `isInvitationFlow` parameter from the `useOnLogin` hook, and chose instead to rely on the existence of the `backRedirectPath` parameter, since it seems that's what `isInvitationFlow` was meant to do anyway.