M2-6042: Validate Params from Admin to Web App (#438)
This PR builds on #431 and introduces validation of the following query parameters:

- `startActivityOrFlow` - An activity or Flow ID
- `respondentId` - The respondent taking the assessment
- `subjectId` - The subject who the assessment is about

If one of these parameter is omitted, then no validation takes place and the app behaves as it does today with one exception: the activity or flow is not started. Starting the activity/flow automatically now requires the provision of all three parameters.

If they are all provided, then the following validation is performed:
- `startActivityOrFlow` identifies a valid activity
- `respondentId` matches the currently authenticated user
- `subjectId` corresponds to a real subject
- The currently authenticated user has one of these roles in the applet
  - Super Admin
  - Owner
  - Manager

If all the validation passes, then the activity/flow is started automatically and the context is stored in a multi-informant redux store.