feat: Updating subject fetching logic (M2-7035) (#506)
* Updating subject fetching logic

I've added a new query that allows both full account participants and team members to fetch their own subject details.

I've also modified the logic for fetching the source and target subjects. In the admin panel, a temporary relation is only created if the subjects differ.
Since this relation is required for fetching the subjects, I conditionally fetch them based on this difference.

* Fix query enabled checks

Tanstack query depends on a proper boolean value (not boolean | undefined) to disable the query

* Simplify boolean checks

At this point in the code we've already narrowed the type, so we don't have to worry about `undefined`

* Change isLoading to isFetching

According to tanstack query docs, isLoading will be true even if th query is disabled, so we should use `fetchStatus` to determine if it is actively fetching data. `isFetching` is a convenience property for this.

See https://tanstack.com/query/v4/docs/framework/react/reference/useQuery