feat: per-subject activity progress/submission
Made widespread change to incorporate `targetSubjectId` into activity
progres tracking, when assignments feature is enabled. Anytime `eventId`
is used, now `targetSubjectid` also needs to be passed to identify which
activity assignment is the active assessment. This value is also passed
to the `answers` POST endpoint to ensure the submission is also properly
associated with the `targetSubjectId`.

For self-reports, and when assignments feature is disabled,
`targetSubjectId` is set to `null`, effectively inheriting existing
functionality.

For `PublicSurvey` and `PublicAutoCompletion`, `targetSubjectId` stays
`null` since public assessments do not support activity assignments.