feat: add subject label to item descriptions
This change required a markdown parser to be able to inject the label
below any paragraphs that contain only media (images, videos, audio)
and above the first paragraph that contains textual content.

Also noticed that the `Banners` component was incorrectly placed in a
scrollable container, making banner alerts (such as validation errors)
not appear to users who have scrolled down on an item that has a long
description.