feat: add SSI props to assessment starting events
Create Mixpanel utilities for conveniently adding the `Item Types` and
optionally populating the `Feature` property for all events tracked
during an assessment. Simplifies some existing tracking code.

The the `Feature` property is now an array after discussions with
the data team, since it's now used for multiple features that can be
present at the same time, and as an array it requires careful handling.