chore: Mixpanel Refactor
This PR modifies the way events are tracked with Mixpanel to match a [similar change](https://github.com/ChildMindInstitute/mindlogger-admin/pull/1959) in the admin repo. There is now a central place to define events and their properties so that we can conveniently see which properties are included in each event, and where they are used.