feat: add SSI properties to "Assessment completed"
To be able to add "Item Types" to this and other events required of this
Jira task, a call to the `applets/{applet_id}/base_info` endpoint is
needed to fetch the item types contained in the activity being
completed. As anoying as it is to have to make another request because
of incomplete data coming from `applets/{applet_id}`, this seems to be
the most convenient way forward for now. We can optimize performance if
needed down the road; it will likely require fixing the BE to return
more consistent data.