fix: normalize self-report targetSubjectID
In the web app, self-report assessments are identified and tracked using
`targetSubjectId=null`. However, for past submissions, the BE returns
the raw value of `targetSubjectId`, which is the subject ID of the
respondent. These submissions needed to be normalized to have their
`targetSubjectId` set to `null`, but to do so, needed to find out what
the subject ID of the respondent (currently logged in user) is.

So added a new `respondentMeta` property to the applet `base_info`
endpoint on the BE, so that this value is now available to the FE to
perform the needed equality comparison against `targetSubjectId`.