fix: use local time zone for existing submissions
Depending on your time zone and the time of day, the date returned by
the submissions endpoint could have been parsed incorrectly using the
previous approach. As a result, previous same-day submissions were
being interpreted as having occurred on the previous day and resulting
in daily activities showing when they shouldn't be.

Fixed this using a simpler and time zone-safe parsing method.