feat: Mixpanel Event Tracking – SSI/Action Plan (M2-7698) (#543)
* feat: add SSI props to assessment starting events

Create Mixpanel utilities for conveniently adding the `Item Types` and
optionally populating the `Feature` property for all events tracked
during an assessment. Simplifies some existing tracking code.

The the `Feature` property is now an array after discussions with
the data team, since it's now used for multiple features that can be
present at the same time, and as an array it requires careful handling.

* feat: add SSI properties to "Assessment completed"

To be able to add "Item Types" to this and other events required of this
Jira task, a call to the `applets/{applet_id}/base_info` endpoint is
needed to fetch the item types contained in the activity being
completed. As anoying as it is to have to make another request because
of incomplete data coming from `applets/{applet_id}`, this seems to be
the most convenient way forward for now. We can optimize performance if
needed down the road; it will likely require fixing the BE to return
more consistent data.

* feat: track new events required by task

Tracks `Report Generated`, `Report Download Clicked` and
`Save & Exit Clicked` events.

* test: add tests for new utilities