fix: Hide one-time completion self-report assessments after complete (M2-8164, M2-8026) (#547)
* fix: normalize self-report targetSubjectID

In the web app, self-report assessments are identified and tracked using
`targetSubjectId=null`. However, for past submissions, the BE returns
the raw value of `targetSubjectId`, which is the subject ID of the
respondent. These submissions needed to be normalized to have their
`targetSubjectId` set to `null`, but to do so, needed to find out what
the subject ID of the respondent (currently logged in user) is.

So added a new `respondentMeta` property to the applet `base_info`
endpoint on the BE, so that this value is now available to the FE to
perform the needed equality comparison against `targetSubjectId`.

* fix: use local time zone for existing submissions

Depending on your time zone and the time of day, the date returned by
the submissions endpoint could have been parsed incorrectly using the
previous approach. As a result, previous same-day submissions were
being interpreted as having occurred on the previous day and resulting
in daily activities showing when they shouldn't be.

Fixed this using a simpler and time zone-safe parsing method.