fix: Link event with in-progress activities
Save the scheduled event that an activity was started with to its
`ActivityOrFlowProgress` object, which is persisted along with the
in-progress state. Then use this persisted event's version when
submitting answers, rather the most current version of the event.

Within the context of performing an assessment, it's also important to
fetch the `timers` portion of the event from the persisted object,
rather than `SurveyContext` as was being done prior.

To ensure the timer state is shown accurately for in-progress activities
or flows listed in the `ActivityGroupList`, it required some minor
refactoring of `ActivityGroupsBuildManager`:
- First, we iterate through the list of in-progress activities/flows as
derived from persisted state and use the events associated with them.
- Then, we do the same original iteration through the events available
on the BE to derive the contents of the "Available" and "Scheduled"
groups of activities/flows.