feat: `/active-assessment` nav to EHR last substep
When the current assessment is being opened via the `/active-assessment`
route, check if there's an EHR in this activity, and the EHR is the
currently active step, and the 1UpHealth iframe is the currently active
substep. If so, skip to the last step, because it's presumed the only
time this will happen is when this route is called by the 3rd-party
health data provider, under those specific circumstances.