feat: Generate `submit_id` when starting surveys
This change renames the `executionGroupKey` property, which used to only
be present on activity flow assessments, to `submitId` (since that's its
real purpose), and also attaches this property to regular activity
assessments. Generating it when starting (or restarting) any assessment
ensures it's available at any time during the assessment, and thus can
be passed as a parameter to the 1UpHealth token endpoint.