refactor: Make back/next visibility more flexible
Based on current discussions in Figma we may end up wanting more
granular control of back & next button visibility on certain sub-steps,
so adjusted return value of `useSubSteps` hook to support that.

Also moved the mutation to reset the Yes/No selection on the additional
EHRs step to the screen itself, so that the previous answer to that
question is preserved in support of adjusting nav logic of the 1UpHealth
sub-step based on that selection.