feat: Add full 1UpHealth respondent flow for consenting to share EHRs (M2-8849) (#616)
* feat: Add `/active-assessment` route

Keep track of the activity that was most recently open by storing its
group progress ID, along with applet ID and optional public applet key.
When visiting this new route, the activity is resumed at where it was
left off.

* feat: Add support for multi-step items

Also add preliminary EHR item type-specific handling of substep
navigation to new hook, `useSubSteps`.

Added some missing DTO types as well.

* refactor: Add `ItemMarkdown` component

We need to reuse the functionality to insert the multi-informant badge
above markdown content, and the "Optional" text below the markdown,
inside `RequestHealthRecordDataItem` and `ActivityCardItem` components.
Created a new component `ItemMarkdown` to be reused in both places.

Slightly refactored `RequestHealthRecordDataItem` to include the
external link inside the unprocessed markdown so that the "Optional"
label is displayed below it.

* feat: Flesh out substep screens of EHR item type

Added support for saving custom properties to specific event types, in
this case `additionalEHRs`. This value, which represents the choice the
respondent made on the last substep, will inform how navigation happens
between substeps.

* feat: Add substep navigation logic

* feat: `/active-assessment` nav to EHR last substep

When the current assessment is being opened via the `/active-assessment`
route, check if there's an EHR in this activity, and the EHR is the
currently active step, and the 1UpHealth iframe is the currently active
substep. If so, skip to the last step, because it's presumed the only
time this will happen is when this route is called by the 3rd-party
health data provider, under those specific circumstances.

* refactor: Split EHR item into substep components

* feat: Add 1UpHealth iframe & endpoint

* feat: Generate `submit_id` when starting surveys

This change renames the `executionGroupKey` property, which used to only
be present on activity flow assessments, to `submitId` (since that's its
real purpose), and also attaches this property to regular activity
assessments. Generating it when starting (or restarting) any assessment
ensures it's available at any time during the assessment, and thus can
be passed as a parameter to the 1UpHealth token endpoint.

* style: Mobile/styling fixes

* fix: TS fixes

* test: Add tests for `NavigateToActiveAssessment`

Adding these tests required the ability to provide a mock dispatch
function and disable the router in the `renderWithProviders` utility.

* fix: Change layout to left-aligned, add MI badge

After chatting with Geoff, it was agreed that the sub-screens should all
have left-aligned text to better fit the styling of the multi-informant
badge, which he requested be present on the Partnerhip & Additional EHRs
prompt screens.

* refactor: Make back/next visibility more flexible

Based on current discussions in Figma we may end up wanting more
granular control of back & next button visibility on certain sub-steps,
so adjusted return value of `useSubSteps` hook to support that.

Also moved the mutation to reset the Yes/No selection on the additional
EHRs step to the screen itself, so that the previous answer to that
question is preserved in support of adjusting nav logic of the 1UpHealth
sub-step based on that selection.

* test: Add tests for `useSubSteps`

* fix: Update sub-step navigation logic

It turns out we will not be hiding either the "Back" or "Next" buttons
on any of the sub-steps after all. Their behaviour is just tweaked
slightly based on the current sub-step and the respondent's responses.
On the OneUpHealth step, the "Next" button should have its text changed
to "Skip", and it skips the user to the next item in the activity.

* fix: Make final change to navigation logic

It was requested that after pressing "Back" on the OneUpHealth step that
the user should be taken back to the AdditionalPrompt step if they had
arrived there after having selected "Yes" on that step. That meant the
default additional EHRs prompt response needed to be reset only when
arriving there via the `/active-assessment` route, rather than every
time you arrive at that screen.

Also, to animate these steps correctly, the animation logic had to be
slightly hijacked.

Unrelatedly, corrected the `RequestHealthRecordDataItemConfigDTO` type
definition (actually has no impact on anything).

* fix: PR review fixes

Also increased the number of retries for receiving the 1UpHealth access
token to 4 (5 total attempts) since it appears that sometimes 1UpHealth
returns an error state (unclear why this may happen, thus far – will
keep an eye on it).

* fix: PR review suggestions

* fix: Convert opt-in/opt-out strings to enum