feat: Add EHR redirect interstitial page for desktop and mobile (M2-9361) (#640)
This PR adds an interstitial page at the path `/ehr-complete` to act as the redirect path for 1UpHealth apps. Users who start an EHR in the mobile app on Android will eventually get redirected here and will be required to click a button to trigger the re-opening of the app.

The motivation for this is that Android apps require an explicit user interaction in order to intercept a deep link. 1UpHealth uses an automatic client side redirect that lacks user interaction, so the Android app fails to open.